server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Roman Vlasenko <klavionik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.40"}
httpx = "^0.23.0"
asyncpg = "^0.26.0"
alembic = "^1.8.1"
pydantic = {extras = ["dotenv"], version = "^1.9.2"}
uvicorn = "^0.18.3"
ptpython = "^3.0.20"
fastapi = "^0.82.0"
itsdangerous = "^2.1.2"
PyJWT = {extras = ["cryptography"], version = "^2.4.0"}
ormar = "^0.11.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
